refactor : 각 db table의 name 필드가 unique로 선언되어있어, station/line의 name에 대한 중복 검사를 제거.
